Extract the hnsw graph merging from being part of the vector writer
