+++ lucene/core/src/java/org/apache/lucene/util/hnsw/HnswGraphBuilder.java
@@ -18,14 +18,11 @@
 package org.apache.lucene.util.hnsw;
 
 import static java.lang.Math.log;
-import static org.apache.lucene.search.DocIdSetIterator.NO_MORE_DOCS;
 
 import java.io.IOException;
-import java.util.HashSet;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
 import java.util.SplittableRandom;
 import java.util.concurrent.TimeUnit;
 import org.apache.lucene.search.KnnCollector;
@@ -37,7 +34,7 @@ import org.apache.lucene.util.InfoStream;
  * Builder for HNSW graph. See {@link HnswGraph} for a gloss on the algorithm and the meaning of the
  * hyper-parameters.
  */
-public final class HnswGraphBuilder {
+public class HnswGraphBuilder {
 
   /** Default number of maximum connections per node */
   public static final int DEFAULT_MAX_CONN = 16;
@@ -67,11 +64,9 @@ public final class HnswGraphBuilder {
   private final GraphBuilderKnnCollector
       beamCandidates; // for levels of graph where we add the node
 
-  final OnHeapHnswGraph hnsw;
+  protected final OnHeapHnswGraph hnsw;
 
   private InfoStream infoStream = InfoStream.getDefault();
-
-  private final Set<Integer> initializedNodes;
 
   public static HnswGraphBuilder create(
       RandomVectorScorerSupplier scorerSupplier, int M, int beamWidth, long seed)
@@ -87,9 +82,8 @@ public final class HnswGraphBuilder {
       HnswGraph initializerGraph,
       Map<Integer, Integer> oldToNewOrdinalMap)
       throws IOException {
-    HnswGraphBuilder hnswGraphBuilder = new HnswGraphBuilder(scorerSupplier, M, beamWidth, seed);
-    hnswGraphBuilder.initializeFromGraph(initializerGraph, oldToNewOrdinalMap);
-    return hnswGraphBuilder;
+    return new InitializedHnswGraphBuilder(
+        scorerSupplier, M, beamWidth, seed, initializerGraph, oldToNewOrdinalMap);
   }
 
   /**
@@ -103,7 +97,7 @@ public final class HnswGraphBuilder {
    * @param seed the seed for a random number generator used during graph construction. Provide this
    *     to ensure repeatable construction.
    */
-  private HnswGraphBuilder(
+  protected HnswGraphBuilder(
       RandomVectorScorerSupplier scorerSupplier, int M, int beamWidth, long seed)
       throws IOException {
     if (M <= 0) {
@@ -126,7 +120,6 @@ public final class HnswGraphBuilder {
     scratch = new NeighborArray(Math.max(beamWidth, M + 1), false);
     entryCandidates = new GraphBuilderKnnCollector(1);
     beamCandidates = new GraphBuilderKnnCollector(beamWidth);
-    this.initializedNodes = new HashSet<>();
   }
 
   /**
@@ -142,45 +135,6 @@ public final class HnswGraphBuilder {
     return hnsw;
   }
 
-  /**
-   * Initializes the graph of this builder. Transfers the nodes and their neighbors from the
-   * initializer graph into the graph being produced by this builder, mapping ordinals from the
-   * initializer graph to their new ordinals in this builder's graph. The builder's graph must be
-   * empty before calling this method.
-   *
-   * @param initializerGraph graph used for initialization
-   * @param oldToNewOrdinalMap map for converting from ordinals in the initializerGraph to this
-   *     builder's graph
-   */
-  private void initializeFromGraph(
-      HnswGraph initializerGraph, Map<Integer, Integer> oldToNewOrdinalMap) throws IOException {
-    assert hnsw.size() == 0;
-    for (int level = 0; level < initializerGraph.numLevels(); level++) {
-      HnswGraph.NodesIterator it = initializerGraph.getNodesOnLevel(level);
-
-      while (it.hasNext()) {
-        int oldOrd = it.nextInt();
-        int newOrd = oldToNewOrdinalMap.get(oldOrd);
-
-        hnsw.addNode(level, newOrd);
-
-        if (level == 0) {
-          initializedNodes.add(newOrd);
-        }
-
-        NeighborArray newNeighbors = this.hnsw.getNeighbors(level, newOrd);
-        initializerGraph.seek(level, oldOrd);
-        for (int oldNeighbor = initializerGraph.nextNeighbor();
-            oldNeighbor != NO_MORE_DOCS;
-            oldNeighbor = initializerGraph.nextNeighbor()) {
-          int newNeighbor = oldToNewOrdinalMap.get(oldNeighbor);
-          // we will compute these scores later when we need to pop out the non-diverse nodes
-          newNeighbors.addOutOfOrder(newNeighbor, Float.NaN);
-        }
-      }
-    }
-  }
-
   /** Set info-stream to output debugging information * */
   public void setInfoStream(InfoStream infoStream) {
     this.infoStream = infoStream;
@@ -193,9 +147,6 @@ public final class HnswGraphBuilder {
   private void addVectors(int maxOrd) throws IOException {
     long start = System.nanoTime(), t = start;
     for (int node = 0; node < maxOrd; node++) {
-      if (initializedNodes.contains(node)) {
-        continue;
-      }
       addGraphNode(node);
       if ((node % 10000 == 0) && infoStream.isEnabled(HNSW_COMPONENT)) {
         t = printGraphBuildStatus(node, start, t);
